requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
certifi==2025.11.12
chamfer==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
emd_ext==0.0.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scipy==1.15.3
sentry-sdk==2.44.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.22
tokenizers==0.19.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.67.1
transformers==4.41.2
triton==2.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.23.0
Werkzeug==3.1.3
